Implemented recommended addons feature, updated readme, fixed theme settings
Added the ability to download recommended addons. The application will check if winget is installed, if it's not, it will download it for you and install the applications using winget. You may also select the download bundle option which will download all of the applications listed in the addons. Bugs were fixed where the theme wasn't applying the custom wallpaper and the readme was updated